+++ src/main/java/com/dynatrace/openkit/core/objects/SessionImpl.java
@@ -121,6 +121,22 @@ public class SessionImpl extends OpenKitComposite implements Session {
     }
 
     @Override
+    public void reportCrash(Throwable throwable) {
+        if (throwable == null) {
+            logger.warning(this + "reportCrash: throwable must not be null");
+            return;
+        }
+        if (logger.isDebugEnabled()) {
+            logger.debug(this + "reportCrash(" + throwable + ")");
+        }
+        synchronized (state) {
+            if (!state.isFinishingOrFinished()) {
+                beacon.reportCrash(throwable);
+            }
+        }
+    }
+
+    @Override
     public WebRequestTracer traceWebRequest(URLConnection connection) {
         if (connection == null) {
             logger.warning(this + "traceWebRequest (URLConnection): connection must not be null");